DonorCheck v1.2.15

+++ src/main/java/org/pankratzlab/unet/deprecated/hla/AntigenDictionary.java
@@ -34,7 +34,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.Callable;
-import java.util.stream.Collectors;
 import org.pankratzlab.unet.deprecated.util.SerializeUtils;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableSet;
@@ -49,7 +48,7 @@ public final class AntigenDictionary implements Serializable {
   public static final String REL_DNA_SER_PROP = "rel.dna.ser.file";
 
   public static final String SERIALIZED_MAP = Info.HLA_HOME + ".hla/map.ser";
-  private static final String MASTER_MAP_RECORDS = "rel_dna_ser.txt";
+  public static final String MASTER_MAP_RECORDS = "rel_dna_ser.txt";
   private static final String COMMENT = "#";
   private static final String COL_DELIM = ";";
   private static final String SPEC_DELIM = ":";
@@ -267,13 +266,14 @@ public final class AntigenDictionary implements Serializable {
       SerializeUtils.write(typeMap, SERIALIZED_MAP);
       map = typeMap;
 
-      List<HLAType> multi = map.hlaDict.keySet().stream()
-          .filter(ht -> map.hlaDict.get(ht).size() > 1).collect(Collectors.toList());
-      System.out.println("Found " + multi.size() + " HLAType(s) with multiple SeroType mappings:");
-      for (HLAType t : multi) {
-        System.out.println("\t" + t.toString() + " --> "
-            + map.hlaDict.get(t).stream().map(s -> s.toString()).collect(Collectors.joining(", ")));
-      }
+      // List<HLAType> multi = map.hlaDict.keySet().stream()
+      // .filter(ht -> map.hlaDict.get(ht).size() > 1).collect(Collectors.toList());
+      // System.out.println("Found " + multi.size() + " HLAType(s) with multiple SeroType
+      // mappings:");
+      // for (HLAType t : multi) {
+      // System.out.println("\t" + t.toString() + " --> "
+      // + map.hlaDict.get(t).stream().map(s -> s.toString()).collect(Collectors.joining(", ")));
+      // }
 
     } catch (Exception e) {
       e.printStackTrace();
@@ -298,7 +298,7 @@ public final class AntigenDictionary implements Serializable {
     return getVersion(filePath);
   }
 
-  private static String getVersion(String file) {
+  public static String getVersion(String file) {
     try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
       return findVersion(reader);
     } catch (Throwable e) {
@@ -307,7 +307,7 @@ public final class AntigenDictionary implements Serializable {
     return null;
   }
 
-  private static String getBundledVersion() {
+  public static String getBundledVersion() {
     try (BufferedReader reader = new BufferedReader(new InputStreamReader(
         AntigenDictionary.class.getClassLoader().getResourceAsStream(MASTER_MAP_RECORDS)))) {
       return findVersion(reader);
@@ -329,7 +329,9 @@ public final class AntigenDictionary implements Serializable {
 
   public static void clearCache() {
     try {
-      Files.delete(Paths.get(SERIALIZED_MAP));
+      if (new File(SERIALIZED_MAP).exists()) {
+        Files.delete(Paths.get(SERIALIZED_MAP));
+      }
       map = null;
     } catch (IOException e) {
       throw new RuntimeException(
